+++ integration-tests-spring-boot/src/main/java/module-info.java
@@ -1,3 +1,5 @@
+import io.github.tomdw.java.modules.spring.samples.basicapplication.speaker.api.SpeakerService;
+
 open module io.github.tomdw.java.modules.spring.integration.tests.spring.boot {
 	requires spring.boot.autoconfigure;
 	requires spring.context;
@@ -6,4 +8,7 @@ open module io.github.tomdw.java.modules.spring.integration.tests.spring.boot {
 	requires io.github.tomdw.java.modules.context.boot;
 
 	requires io.github.tomdw.java.modules.spring.samples.basicapplication.application;
+	requires io.github.tomdw.java.modules.spring.samples.basicapplication.speaker;
+
+	uses SpeakerService;
 }