+++ integration-tests/src/main/java/module-info.java
@@ -1,11 +1,11 @@
 import io.github.tomdw.java.modules.context.boot.api.ModuleContext;
-import io.github.tomdw.java.modules.spring.integration.tests.IntegrationTestService;
+import io.github.tomdw.java.modules.spring.integration.tests.IntegrationTestConfiguration;
 import io.github.tomdw.java.modules.spring.samples.basicapplication.speaker.api.MultipleSpeakerService;
 import io.github.tomdw.java.modules.spring.samples.basicapplication.speaker.api.MultipleSpeakerWithGenericsService;
 import io.github.tomdw.java.modules.spring.samples.basicapplication.speaker.api.NamedSpeakerService;
 import io.github.tomdw.java.modules.spring.samples.basicapplication.speaker.api.SpeakerService;
 
-@ModuleContext(mainConfigurationClass = IntegrationTestService.class)
+@ModuleContext(mainConfigurationClass = IntegrationTestConfiguration.class)
 module io.github.tomdw.java.modules.spring.integration.tests {
 	requires io.github.tomdw.java.modules.context.boot;
 	requires io.github.tomdw.java.modules.spring.samples.basicapplication.application;
@@ -16,5 +16,5 @@ module io.github.tomdw.java.modules.spring.integration.tests {
 	uses NamedSpeakerService;
 	uses MultipleSpeakerWithGenericsService;
 
-	opens io.github.tomdw.java.modules.spring.integration.tests to spring.beans, spring.core;
+	opens io.github.tomdw.java.modules.spring.integration.tests to spring.beans, spring.core, spring.context;
 }