+++ java-modules-context-boot/src/main/java/io/github/tomdw/java/modules/context/boot/internal/ModuleServiceReferenceAnnotationPostProcessor.java
@@ -3,10 +3,9 @@ package io.github.tomdw.java.modules.context.boot.internal;
 import static java.lang.System.Logger.Level.ERROR;
 import static java.lang.System.Logger.Level.INFO;
 
+import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
-import java.lang.reflect.ParameterizedType;
-import java.util.List;
-import java.util.Optional;
+import java.lang.reflect.Parameter;
 
 import javax.inject.Named;
 
@@ -14,7 +13,6 @@ import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.BeanCreationException;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.BeanFactoryAware;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.beans.factory.config.InstantiationAwareBeanPostProcessorAdapter;
 import org.springframework.beans.factory.serviceloader.AbstractServiceLoaderBasedFactoryBean;
@@ -36,61 +34,45 @@ public class ModuleServiceReferenceAnnotationPostProcessor extends Instantiation
 		Module usingModule = beanClass.getModule();
 		for (Field field : beanClass.getDeclaredFields()) {
 			if (field.isAnnotationPresent(ModuleServiceReference.class)) {
-				Class<?> injectionType = field.getType();
-				if(injectionType.isAssignableFrom(List.class)) {
-					registerServiceListFactoryBean(usingModule, field);
-				} else {
-					registerServiceFactoryBean(usingModule, field);
+				postProcessModuleServiceReference(usingModule, ModuleServiceReferenceElement.of(field));
+			}
+		}
+		for (Constructor<?> constructor: beanClass.getDeclaredConstructors()) {
+			for(Parameter constructorParameter: constructor.getParameters()) {
+				if (constructorParameter.isAnnotationPresent(ModuleServiceReference.class)) {
+					postProcessModuleServiceReference(usingModule, ModuleServiceReferenceElement.of(constructorParameter));
 				}
 			}
 		}
 		return null;
 	}
 
-	private void registerServiceFactoryBean(Module usingModule, Field field) {
-		Class<?> serviceType = field.getType();
+	private void postProcessModuleServiceReference(Module usingModule, ModuleServiceReferenceElement serviceReferenceElement) {
+		if(serviceReferenceElement.isServicesList()) {
+			registerServiceListFactoryBean(usingModule, serviceReferenceElement);
+		} else {
+			registerServiceFactoryBean(usingModule, serviceReferenceElement);
+		}
+	}
+
+	private void registerServiceFactoryBean(Module usingModule, ModuleServiceReferenceElement serviceReferenceElement) {
+		Class<?> serviceType = serviceReferenceElement.getModuleServiceType();
 		LOGGER.log(INFO, "ModuleServiceReference from module " + usingModule.getName() + " found for service of type " + serviceType);
-		String newBeanName = serviceFactoryBeanName(serviceType);
 		validateModuleUsesServiceType(usingModule, serviceType);
+		String newBeanName = serviceReferenceElement.getBeanNameToUseForServiceFactory();
 		if(!this.beanFactory.containsBean(newBeanName)) {
 			registerAbstractServiceFactoryBean(serviceType, newBeanName, new ServiceFactoryBean());
 		}
 	}
 
-	private void registerServiceListFactoryBean(Module usingModule, Field field) {
-		Class<?> serviceType = extractTypeOfField(field);
+	private void registerServiceListFactoryBean(Module usingModule, ModuleServiceReferenceElement serviceReferenceElement) {
+		Class<?> serviceType = serviceReferenceElement.getModuleServiceType();
 		LOGGER.log(INFO, "ModuleServiceReference from module " + usingModule.getName() + " found for service of type List of " + serviceType);
 		validateModuleUsesServiceType(usingModule, serviceType);
-		validateNamedOrQualifierAnnotationPresent(field, serviceType);
-		String newBeanName = serviceListFactoryBeanName(serviceType);
+		String newBeanName = serviceReferenceElement.getBeanNameToUseForServiceFactory();
 		if (!this.beanFactory.containsBean(newBeanName)) {
 			registerAbstractServiceFactoryBean(serviceType, newBeanName, new ServiceListFactoryBean());
 		}
-	}
-
-	private Class<?> extractTypeOfField(Field field) {
-		ParameterizedType listGenericType = (ParameterizedType) field.getGenericType();
-		if (checkIfContentOfGenericListHasGenerics(listGenericType)) {
-			return (Class<?>) ((ParameterizedType) listGenericType.getActualTypeArguments()[0]).getRawType();
-		} else {
-			return (Class<?>) listGenericType.getActualTypeArguments()[0];
-		}
-	}
-
-	private boolean checkIfContentOfGenericListHasGenerics(ParameterizedType listGenericType) {
-		return listGenericType.getActualTypeArguments()[0] instanceof ParameterizedType;
-	}
-
-	private String serviceFactoryBeanName(Class<?> serviceType) {
-		return decapitalizeFirstLetter(serviceType.getSimpleName());
-	}
-
-	private String serviceListFactoryBeanName(Class<?> serviceType) {
-		return decapitalizeFirstLetter(serviceType.getSimpleName() + "List");
-	}
-
-	private String decapitalizeFirstLetter(String input) {
-		return Character.toLowerCase(input.charAt(0)) + input.substring(1);
 	}
 
 	private void registerAbstractServiceFactoryBean(Class<?> serviceType, String newBeanName, AbstractServiceLoaderBasedFactoryBean serviceFactoryBean) {
@@ -108,23 +90,6 @@ public class ModuleServiceReferenceAnnotationPostProcessor extends Instantiation
 			LOGGER.log(ERROR, "Module with name " + usingModule.getName() + " does not define a 'use' dependency on service of type: use " + serviceType.getName());
 			throw new IllegalStateException("Module with name " + usingModule.getName() + " does not define a 'use' dependency on service of type: use " + serviceType.getName());
 		}
-	}
-
-	private void validateNamedOrQualifierAnnotationPresent(Field field, Class<?> serviceType) {
-		if(getNamedOrQualifierValue(field).isEmpty() || !getNamedOrQualifierValue(field).orElseThrow().equals(serviceListFactoryBeanName(serviceType))) {
-			LOGGER.log(ERROR, "Injecting a list of " + serviceType.getName() + " requires a Qualifier or Named annotation with value " + serviceListFactoryBeanName(serviceType));
-			throw new IllegalStateException("Injecting a list of " + serviceType.getName() + " requires a Qualifier or Named annotation with value " + serviceListFactoryBeanName(serviceType));
-		}
-	}
-
-	private Optional<String> getNamedOrQualifierValue(Field field) {
-		if(field.isAnnotationPresent(Named.class)) {
-			return Optional.of(field.getAnnotation(Named.class).value());
-		}
-		if(field.isAnnotationPresent(Qualifier.class)) {
-			return Optional.of(field.getAnnotation(Qualifier.class).value());
-		}
-		return Optional.empty();
 	}
 
 	@Override